refactor: move rate tracking into provider-neutral package
GitLab rate-limit handling should not depend on the GitHub client or go-github response types. Move the tracker implementation into internal/ratelimit, keep GitHub compatibility constructors, and convert provider API responses at their package boundaries.
